+++ package.nls.json
@@ -3,5 +3,6 @@
   "salesforcedx-vscode-agents.openAgentInOrg.description": "SFDX: Open an Agent in the default Org Picture description",
   "agent_test_view_name": "Agent Tests",
   "run_all_tests_title": "Run All Test Definitions",
-  "run_test_title": "Run Test Definition"
+  "run_test_title": "Run Test Definition",
+  "refresh_test_title": "Refresh Test Definitions"
 }
