+++ src/enums/commands.ts
@@ -1,3 +1,7 @@
 export enum Commands {
-  openAgentInOrg = 'salesforcedx-vscode-agents.openAgentInOrg'
+  openAgentInOrg = 'salesforcedx-vscode-agents.openAgentInOrg',
+  goToDefinition = 'sf.agent.test.view.goToDefinition',
+  runTest = 'sf.agent.test.view.runTest',
+  refreshTestView = 'sf.agent.test.view.refresh',
+  collapseAll = 'sf.agent.test.view.collapseAll'
 }
