+++ src/types/types.ts
@@ -1,0 +1,15 @@
+import vscode from "vscode";
+
+export type AiEvaluationDefinition = {
+    testSetName: string;
+    name: string;
+    subjectName: string;
+    description: string;
+    testCases: AgentTestCase[];
+    location: vscode.Location;
+};
+
+export type AgentTestCase = {
+    location: vscode.Location;
+    number: string;
+};
