+++ tsconfig.tsbuildinfo
@@ -1,1 +1,1 @@
-{"root":["./src/extension.ts","./src/index.ts","./src/commands/index.ts","./src/commands/openagentinorg.ts","./src/enums/commands.ts","./src/views/testoutlineprovider.ts","./src/views/testrunner.ts"],"version":"5.7.2"}
+{"root":["./src/extension.ts","./src/index.ts","./src/commands/index.ts","./src/commands/openagentinorg.ts","./src/enums/commands.ts","./src/types/index.ts","./src/types/testnodes.ts","./src/views/testoutlineprovider.ts","./src/views/testrunner.ts"],"version":"5.7.2"}
